requirements.txt
alpaca-py>=0.30.0
anthropic>=0.39.0
pandas>=2.2.0
ta>=0.11.0
yfinance>=0.2.40
python-dotenv>=1.0.0
sqlalchemy>=2.0.30
python-telegram-bot>=21.0
apscheduler>=3.10.4
aiohttp>=3.9.0
finnhub-python>=2.4.20
